requirements.txt
streamlit
pandas
matplotlib
seaborn
numpy
scipy
scikit-learn
openai
groq
google-generativeai
python-dotenv
